Avoid recursion in `_extract_serialize`
Instead of recursively calling `_extract_serialize`, simply maintain a
queue of collections to visit. When starting out, pop the first
collection from the queue and process it. As we encounter new
collections, add them to the queue to visit later. This way we still
have the behavior as if we were recursing, but we lower the overhead by
not having to track a deep stack with many things that don't need to be
included.